Maximum size for StateChange storage
This switches the key generation to use the instance ID and the
version number to sort the key-value pairs. It also add a cleaning
procedure when appending new items to fit the maximum size. Note that
it can be slightly above, it is not strict.

Fixes #1537

+++ byzcoin/proto.go
@@ -243,7 +243,7 @@ type StateChange struct {
 	// DarcID is the Darc controlling access to this key.
 	DarcID darc.ID
 	// Version is the monotonically increased version of the instance
-	Version int
+	Version uint64
 }
 
 // Coin is a generic structure holding any type of coin. Coins are defined
